third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/popup-unspecified-with-cross-origin.https-expected.txt

This is a testharness.js-based test.
PASS unspecified document opening popup to https://not-web-platform.test:8444 with COOP: ""
PASS unspecified document opening popup to https://not-web-platform.test:8444 with COOP: "unsafe-none"
PASS unspecified document opening popup to https://not-web-platform.test:8444 with COOP: "jibberish"
PASS unspecified document opening popup to https://not-web-platform.test:8444 with COOP: "same-site"
PASS unspecified document opening popup to https://not-web-platform.test:8444 with COOP: "same-site unsafe-allow-outgoing"
PASS unspecified document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin unsafe-allow-outgoing"
FAIL unspecified document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin" assert_equals: name expected "" but got "unspecified_to_CROSS_ORIGIN_same-origin"
FAIL unspecified document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin-allow-popups" assert_equals: name expected "" but got "unspecified_to_CROSS_ORIGIN_same-origin-allow-popups"
Harness: the test ran to completion.
